requirements.txt
fastf1
matplotlib
numpy
seaborn
scipy
customtkinter
Pillow
